requirements.txt
torch
torchvision
streamlit
pillow
numpy
matplotlib
tqdm
scikit-learn
seaborn
